requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0

# Export and quantization
onnx>=1.15.0
onnxruntime>=1.16.0

# Dataset and tokenization
datasets>=2.14.0
tokenizers>=0.15.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
psutil>=5.9.0

# Optional: benchmarking
memory_profiler>=0.61.0
